start of schema refactor
