+++ api/contact.py
@@ -13,13 +13,9 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 # ===============================================================================
-from typing import List
 
-from fastapi import APIRouter, Depends
-from pydantic import BaseModel
-from pydantic_core import PydanticUndefined
+from fastapi import APIRouter
 from sqlalchemy import select
-from sqlalchemy.orm import Session
 from starlette import status
 
 from api.pagination import CustomPage
@@ -28,24 +24,11 @@ from fastapi_pagination.ext.sqlalchemy import paginate
 from core.dependencies import session_dependency
 from db import ThingContactAssociation, Thing
 from db.contact import Contact, Email, Phone, Address
-from db.engine import get_db_session
-from schemas.response.contact import (
-    EmailResponse,
-    ContactResponse,
-    AddressResponse,
-    PhoneResponse,
-)
-from schemas.create.contact import CreateContact
-from schemas.update.contact import (
-    UpdateContact,
-    UpdateEmail,
-    UpdatePhone,
-    UpdateAddress,
-)
+from schemas_v2.contact import CreateContact, PhoneResponse, EmailResponse, AddressResponse, ContactResponse, \
+    UpdateContact, UpdateEmail, UpdatePhone, UpdateAddress
 from services.crud_helper import model_patcher
 from services.people_helper import add_contact
 from services.query_helper import (
-    simple_all_getter,
     simple_get_by_id,
     paginated_all_getter,
 )
