+++ api/thing.py
@@ -13,10 +13,9 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 # ===============================================================================
-from typing import List, Annotated, Union
 from shapely.geometry import mapping
 from shapely import wkb
-from fastapi import APIRouter, Depends, Query
+from fastapi import APIRouter, Depends
 from fastapi_pagination.ext.sqlalchemy import paginate
 from sqlalchemy import select
 from sqlalchemy.orm import Session
@@ -34,26 +33,16 @@ from db import (
     Location,
 )
 from db.engine import get_db_session
-from db.group import GroupThingAssociation, Group
 from db.thing.thing import ThingIdLink
 from schemas.base_create import CreateSpring
-from schemas.base_get import GetWell
 from schemas.base_responses import SpringResponse
-from schemas.create.thing import CreateThingIdLink, CreateWell, CreateWellScreen
-from schemas.response.location import LocationResponse
-from schemas.response.thing import (
-    WellResponse,
-    WellScreenResponse,
-    ThingResponse,
-    FeatureCollectionResponse,
-    Feature,
-)
-from schemas.update.location import UpdateLocation
-from schemas.update.thing import UpdateThing, UpdateWell
+from schemas_v2.thing import CreateThingIdLink, CreateWell, CreateWellScreen, ThingResponse, WellResponse, \
+    WellScreenResponse, UpdateThing, UpdateWell
+from schemas_v2.location import LocationResponse
+from schema_v2.location import UpdateLocation
 from services.crud_helper import model_patcher
 from services.query_helper import (
     make_query,
-    simple_all_getter,
     simple_get_by_id,
     paginated_all_getter,
 )
