+++ schemas_v2/lexicon.py
@@ -15,6 +15,8 @@
 # ===============================================================================
 from pydantic import BaseModel
 
+from schemas import ORMBaseModel
+
 
 class CreateLexiconTerm(BaseModel):
     """
@@ -53,3 +55,24 @@ class CreateTriple(BaseModel):
 
 
 # ============= EOF =============================================
+class LexiconTermResponse(ORMBaseModel):
+    """
+    Pydantic model for the response of a lexicon term.
+    This model can be extended to include additional fields as needed.
+    """
+
+    term: str
+    definition: str
+    category: str | None = None
+
+
+class LexiconCategoryResponse(ORMBaseModel):
+    """
+    Pydantic model for the response of a lexicon category.
+    This model can be extended to include additional fields as needed.
+    """
+
+    id: int
+    name: str
+    description: str | None = None
+    # terms: list[LexiconTermResponse] | None = None
